# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DiabeCare API",
    "description": "DiabeCare \u2014 Diabetic Patient Medication Follow-Up System API",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://your-production-domain.com/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/medications/patientf/{patient_id}": {
      "get": {
        "tags": [
          "Medications"
        ],
        "summary": "List all medications for a patient (includes taken_today field; completed medications are included \u2014 filter client-side if needed)",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of medication objects with taken_today field"
          }
        }
      }
    },
    "/medications/": {
      "post": {
        "tags": [
          "Medications"
        ],
        "summary": "Prescribe a medication (admin/doctor)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patient_id",
                  "name",
                  "dosage",
                  "frequency"
                ],
                "properties": {
                  "patient_id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string",
                    "example": "Metformin"
                  },
                  "dosage": {
                    "type": "string",
                    "example": "500mg"
                  },
                  "frequency": {
                    "type": "string",
                    "example": "Twice daily"
                  },
                  "time": {
                    "type": "string",
                    "example": "08:00"
                  },
                  "refill_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-08-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medication prescribed"
          },
          "403": {
            "description": "Admin/doctor access required"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
    "/medications/{med_id}": {
      "put": {
        "tags": [
          "Medications"
        ],
        "summary": "Update medication \u2014 update refill date OR mark as completed (admin/doctor)",
        "parameters": [
          {
            "name": "med_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refill_date": {
                    "type": "string",
                    "format": "date",
                    "nullable": true,
                    "description": "Pass null or empty string to clear the refill date",
                    "example": "2025-09-15"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Pass true to mark the medication course as completed (irreversible)",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Medication updated \u2014 returns updated medication object"
          },
          "400": {
            "description": "Invalid refill_date format"
          },
          "403": {
            "description": "Admin/doctor access required"
          },
          "404": {
            "description": "Medication not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Medications"
        ],
        "summary": "Delete medication (admin/doctor)",
        "parameters": [
          {
            "name": "med_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Medication deleted"
          },
          "403": {
            "description": "Admin/doctor access required"
          },
          "404": {
            "description": "Medication not found"
          }
        }
      }
    },
    "/medications/{med_id}/toggle": {
      "put": {
        "tags": [
          "Medications"
        ],
        "summary": "Mark medication taken or missed for today (patient only)",
        "parameters": [
          {
            "name": "med_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "taken"
                ],
                "properties": {
                  "taken": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns taken_today and updated adherence_rate"
          },
          "403": {
            "description": "Patient access only"
          },
          "404": {
            "description": "Medication not found"
          }
        }
      }
    },
...
  },
...
}